Feat: Atualização geral do projeto e separação dos services.

+++ src/main/java/br/com/IgorRafael/todolist/entity/Todo.java
@@ -13,7 +13,7 @@ public class Todo {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	
 	@NotBlank
 	private String nome;
@@ -26,7 +26,7 @@ public class Todo {
 	public Todo() {
 	}
 	
-	public Todo( Long id, @NotBlank String nome, @NotBlank String descricao, boolean realizado, int prioridade) {
+	public Todo(Integer id, @NotBlank String nome, @NotBlank String descricao, boolean realizado, int prioridade) {
 		this.id = id;
 		this.nome = nome;
 	    this.descricao = descricao;
@@ -44,10 +44,10 @@ public class Todo {
 	}
 	
 	
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 	public String getNome() {
